refactor: mock 설정 공용화 및 env 상수 파일 config로 정리
