+++ src/domains/cart/pages/CartClientPage.tsx
@@ -18,7 +18,7 @@ import { MOCK_GET_CART } from '@/mocks/cart.mock';
 import { MOCK_GET_COURSE_DETAIL } from '@/mocks/course.mock';
 import type { CartItemResponse, PreparePaymentResponse } from '../types/cart';
 import { useTossPayment } from '../hooks/useTossPayment';
-import { USE_MOCK } from '@/shared/constants/env';
+import { USE_MOCK } from '@/shared/constants/config';
 
 export function CartClientPage() {
   const searchParams = useSearchParams();
